+++ src/components/Navbar.jsx
@@ -24,30 +24,30 @@ const Navbar = ({ auth }) => {
   };
 
   return (
-    <nav className="w-full fixed top-0 left-0 bg-white shadow-lg text-black h-20">
-      <div className="max-w-7xl h-full mx-auto px-12 flex items-center justify-between">
-        <Link className="tracking-wider text-xl font-bold" to="/">
+    <nav className="fixed top-0 left-0 w-full h-20 text-black bg-white shadow-lg">
+      <div className="flex items-center justify-between h-full px-6 mx-auto max-w-7xl lg:px-12">
+        <Link className="text-xl font-bold tracking-wider" to="/">
           Instagram
         </Link>
         <div className="flex flex-row items-center justify-center gap-20">
           <Link
-            className="flex flex-row justify-center items-center gap-2 cursor-pointer"
+            className="flex flex-row items-center justify-center gap-2 cursor-pointer"
             to={`/profile/${user.id}`}
           >
             <img
               src={user?.profilePictureUrl || profile}
               alt=""
-              className="w-12 h-12 object-cover rounded-full border-2 border-gray-200"
+              className="object-cover w-12 h-12 border-2 border-gray-200 rounded-full"
             />
             <div>
               <p className="text-base font-bold">{user?.username}</p>
-              <p className="text-sm text-gray-500 font-semibold">
+              <p className="text-sm font-semibold text-gray-500">
                 {user?.name}
               </p>
             </div>
           </Link>
           <button
-            className="bg-gray-400 rounded-full p-3 hover:bg-gray-500 transition-all"
+            className="p-3 transition-all bg-gray-400 rounded-full hover:bg-gray-500"
             onClick={handleLogout}
           >
             <FaSignOutAlt />
