+++ src/components/Posts.jsx
@@ -4,7 +4,7 @@ import ListPost from "./ListPost";
 const Posts = ({ explorePost, explorePage }) => {
   return (
     <section className="mt-24">
-      <div className="flex flex-col items-center w-full px-12 mx-auto max-w-7xl">
+      <div className="flex flex-col items-center w-full px-6 mx-auto lg:px-12 max-w-7xl">
         {explorePost?.map(
           (explore) =>
             explore.user && <ListPost explore={explore} key={explore.id} />
