+++ src/components/index.js
@@ -4,5 +4,26 @@ import InputField from "./InputField";
 import Navbar from "./Navbar";
 import Posts from "./Posts";
 import Story from "./Story";
+import Like from "./Like";
+import ListPost from "./ListPost";
+import Modal from "./Modal";
+import ProfileHeader from "./ProfileHeader";
+import ListFollowers from "./ListFollowers";
+import ListFollowing from "./ListFollowing";
+import UpdateProfileForm from "./UpdateProfileForm";
 
-export { Button, Comments, InputField, Navbar, Posts, Story };
+export {
+  Button,
+  Comments,
+  InputField,
+  Navbar,
+  Posts,
+  Story,
+  Like,
+  ListPost,
+  Modal,
+  ProfileHeader,
+  ListFollowers,
+  ListFollowing,
+  UpdateProfileForm,
+};
